CSHARP-3218: Code review.
